gpu: Restructure gpu folder
