+++ BUILD.gn
@@ -33,7 +33,7 @@ action("vulkan_clean_old_validation_layer_objects") {
   script = "scripts/gn/remove_files.py"
 
   # inputs is a (random) new file since the vvl roll, used to ensure the cleanup is done only once
-  inputs = [ "layers/gpu/gpu_validation.cpp" ]
+  inputs = [ "layers/gpu/core/gpuav.cpp" ]
   outputs = [ "$vulkan_gen_dir/old_vvl_files_are_removed" ]
   args = [
     "$raw_vulkan_gen_dir/old_vvl_files_are_removed",
@@ -163,36 +163,34 @@ vvl_sources = [
   "layers/external/vma/vk_mem_alloc.h",
   "layers/external/vma/vma.cpp",
   "layers/external/vma/vma.h",
-  "layers/gpu_shaders/gpu_error_codes.h",
-  "layers/gpu_shaders/gpu_error_header.h",
-  "layers/gpu_shaders/gpu_shaders_constants.h",
-  "layers/gpu/debug_printf.cpp",
-  "layers/gpu/debug_printf.h",
-  "layers/gpu/gpu_constants.h",
-  "layers/gpu/gpu_descriptor_set.cpp",
-  "layers/gpu/gpu_descriptor_set.h",
-  "layers/gpu/gpu_error_message.cpp",
-  "layers/gpu/gpu_error_message.h",
-  "layers/gpu/gpu_image_layout.cpp",
-  "layers/gpu/gpu_settings.h",
-  "layers/gpu/gpu_shader_instrumentor.cpp",
-  "layers/gpu/gpu_shader_instrumentor.h",
-  "layers/gpu/gpu_state_tracker.cpp",
-  "layers/gpu/gpu_state_tracker.h",
-  "layers/gpu/gpu_subclasses.cpp",
-  "layers/gpu/gpu_subclasses.h",
-  "layers/gpu/gpu_record.cpp",
-  "layers/gpu/gpu_resources.h",
-  "layers/gpu/gpu_settings.h",
-  "layers/gpu/gpu_setup.cpp",
-  "layers/gpu/gpu_validation.cpp",
+  "layers/gpu/core/gpu_settings.h",
+  "layers/gpu/core/gpuav_constants.h",
+  "layers/gpu/cmd_validation/gpuav_cmd_validation_common.h",
+  "layers/gpu/cmd_validation/gpuav_cmd_validation_common.cpp",
   "layers/gpu/cmd_validation/gpuav_draw.cpp",
   "layers/gpu/cmd_validation/gpuav_dispatch.cpp",
   "layers/gpu/cmd_validation/gpuav_trace_rays.cpp",
   "layers/gpu/cmd_validation/gpuav_copy_buffer_to_image.cpp",
-  "layers/gpu/gpu_validation.h",
-  "layers/gpu/gpu_vuids.cpp",
-  "layers/gpu/gpu_vuids.h",
+  "layers/gpu/core/gpuav.h",
+  "layers/gpu/core/gpuav.cpp",
+  "layers/gpu/core/gpuav_state_tracker.cpp",
+  "layers/gpu/core/gpuav_state_tracker.h",
+  "layers/gpu/core/gpuav_record.cpp",
+  "layers/gpu/core/gpuav_setup.cpp",
+  "layers/gpu/core/gpuav_image_layout.cpp",
+  "layers/gpu/debug_printf/debug_printf.cpp",
+  "layers/gpu/debug_printf/debug_printf.h",
+  "layers/gpu/error_message/gpuav_error_message.cpp",
+  "layers/gpu/error_message/gpuav_error_message.h",
+  "layers/gpu/error_message/gpuav_vuids.cpp",
+  "layers/gpu/error_message/gpuav_vuids.h",
+  "layers/gpu/instrumentation/gpuav_shader_instrumentor.cpp",
+  "layers/gpu/instrumentation/gpuav_shader_instrumentor.h",
+  "layers/gpu/resources/gpuav_descriptor_set.cpp",
+  "layers/gpu/resources/gpuav_descriptor_set.h",
+  "layers/gpu/resources/gpuav_subclasses.cpp",
+  "layers/gpu/resources/gpuav_subclasses.h",
+  "layers/gpu/resources/gpuav_resources.h",
   "layers/gpu/spirv/function_basic_block.cpp",
   "layers/gpu/spirv/function_basic_block.h",
   "layers/gpu/spirv/instruction.cpp",
@@ -210,6 +208,9 @@ vvl_sources = [
   "layers/gpu/spirv/buffer_device_address_pass.cpp",
   "layers/gpu/spirv/ray_query_pass.h",
   "layers/gpu/spirv/ray_query_pass.cpp",
+  "layers/gpu_shaders/gpu_error_codes.h",
+  "layers/gpu_shaders/gpu_error_header.h",
+  "layers/gpu_shaders/gpu_shaders_constants.h",
   "layers/layer_options.cpp",
   "layers/layer_options.h",
   "layers/object_tracker/object_lifetime_validation.h",
