+++ layers/gpu/core/gpuav.cpp
@@ -23,15 +23,15 @@
 #include "utils/cast_utils.h"
 #include "utils/shader_utils.h"
 #include "utils/hash_util.h"
-#include "gpu/gpu_constants.h"
-#include "gpu/gpu_subclasses.h"
-#include "gpu/gpu_validation.h"
+#include "gpu/core/gpuav_constants.h"
+#include "gpu/core/gpuav.h"
+#include "gpu/resources/gpuav_subclasses.h"
+#include "gpu/spirv/module.h"
 #include "spirv-tools/instrument.hpp"
 #include "spirv-tools/linker.hpp"
 #include "generated/layer_chassis_dispatch.h"
-#include "gpu_vuids.h"
+#include "gpu/error_message/gpuav_vuids.h"
 #include "containers/custom_containers.h"
-#include "spirv/module.h"
 #include "chassis/chassis_modification_state.h"
 #include "state_tracker/render_pass_state.h"
