+++ layers/gpu/core/gpuav_setup.cpp
@@ -23,15 +23,15 @@
 #include "utils/cast_utils.h"
 #include "utils/shader_utils.h"
 #include "utils/hash_util.h"
-#include "gpu/gpu_constants.h"
-#include "gpu/gpu_validation.h"
-#include "gpu/gpu_subclasses.h"
+#include "gpu/core/gpuav_constants.h"
+#include "gpu/error_message/gpuav_vuids.h"
+#include "gpu/core/gpuav.h"
+#include "gpu/resources/gpuav_subclasses.h"
 #include "state_tracker/device_state.h"
 #include "state_tracker/shader_object_state.h"
 #include "spirv-tools/instrument.hpp"
 #include "spirv-tools/linker.hpp"
 #include "generated/layer_chassis_dispatch.h"
-#include "gpu_vuids.h"
 #include "containers/custom_containers.h"
 #include "gpu_shaders/gpu_error_header.h"
 
@@ -274,217 +274,6 @@ void Validator::CreateDevice(const VkDeviceCreateInfo *pCreateInfo, const Locati
 
         vmaUnmapMemory(vmaAllocator, indices_buffer.allocation);
     }
-}
-
-void PreDrawResources::SharedResources::Destroy(Validator &validator) {
-    if (shader_module != VK_NULL_HANDLE) {
-        DispatchDestroyShaderModule(validator.device, shader_module, nullptr);
-        shader_module = VK_NULL_HANDLE;
-    }
-    if (ds_layout != VK_NULL_HANDLE) {
-        DispatchDestroyDescriptorSetLayout(validator.device, ds_layout, nullptr);
-        ds_layout = VK_NULL_HANDLE;
-    }
-    if (pipeline_layout != VK_NULL_HANDLE) {
-        DispatchDestroyPipelineLayout(validator.device, pipeline_layout, nullptr);
-        pipeline_layout = VK_NULL_HANDLE;
-    }
-    auto to_destroy = renderpass_to_pipeline.snapshot();
-    for (auto &entry : to_destroy) {
-        DispatchDestroyPipeline(validator.device, entry.second, nullptr);
-        renderpass_to_pipeline.erase(entry.first);
-    }
-    if (shader_object != VK_NULL_HANDLE) {
-        DispatchDestroyShaderEXT(validator.device, shader_object, nullptr);
-        shader_object = VK_NULL_HANDLE;
-    }
-}
-
-void PreDispatchResources::SharedResources::Destroy(Validator &validator) {
-    if (ds_layout != VK_NULL_HANDLE) {
-        DispatchDestroyDescriptorSetLayout(validator.device, ds_layout, nullptr);
-        ds_layout = VK_NULL_HANDLE;
-    }
-    if (pipeline_layout != VK_NULL_HANDLE) {
-        DispatchDestroyPipelineLayout(validator.device, pipeline_layout, nullptr);
-        pipeline_layout = VK_NULL_HANDLE;
-    }
-    if (pipeline != VK_NULL_HANDLE) {
-        DispatchDestroyPipeline(validator.device, pipeline, nullptr);
-        pipeline = VK_NULL_HANDLE;
-    }
-    if (shader_object != VK_NULL_HANDLE) {
-        DispatchDestroyShaderEXT(validator.device, shader_object, nullptr);
-        shader_object = VK_NULL_HANDLE;
-    }
-}
-
-void PreTraceRaysResources::SharedResources::Destroy(Validator &validator) {
-    if (ds_layout != VK_NULL_HANDLE) {
-        DispatchDestroyDescriptorSetLayout(validator.device, ds_layout, nullptr);
-        ds_layout = VK_NULL_HANDLE;
-    }
-    if (pipeline_layout != VK_NULL_HANDLE) {
-        DispatchDestroyPipelineLayout(validator.device, pipeline_layout, nullptr);
-        pipeline_layout = VK_NULL_HANDLE;
-    }
-    if (pipeline != VK_NULL_HANDLE) {
-        DispatchDestroyPipeline(validator.device, pipeline, nullptr);
-        pipeline = VK_NULL_HANDLE;
-    }
-    if (sbt_buffer != VK_NULL_HANDLE) {
-        vmaDestroyBuffer(validator.vmaAllocator, sbt_buffer, sbt_allocation);
-        sbt_buffer = VK_NULL_HANDLE;
-        sbt_allocation = VK_NULL_HANDLE;
-        sbt_address = 0;
-    }
-    if (sbt_pool) {
-        vmaDestroyPool(validator.vmaAllocator, sbt_pool);
-        sbt_pool = VK_NULL_HANDLE;
-    }
-}
-
-void PreCopyBufferToImageResources::SharedResources::Destroy(Validator &validator) {
-    if (ds_layout != VK_NULL_HANDLE) {
-        DispatchDestroyDescriptorSetLayout(validator.device, ds_layout, nullptr);
-        ds_layout = VK_NULL_HANDLE;
-    }
-    if (pipeline_layout != VK_NULL_HANDLE) {
-        DispatchDestroyPipelineLayout(validator.device, pipeline_layout, nullptr);
-        pipeline_layout = VK_NULL_HANDLE;
-    }
-    if (pipeline != VK_NULL_HANDLE) {
-        DispatchDestroyPipeline(validator.device, pipeline, nullptr);
-        pipeline = VK_NULL_HANDLE;
-    }
-    if (copy_regions_pool != VK_NULL_HANDLE) {
-        vmaDestroyPool(validator.vmaAllocator, copy_regions_pool);
-        copy_regions_pool = VK_NULL_HANDLE;
-    }
-}
-
-void RestorablePipelineState::Create(vvl::CommandBuffer &cb_state, VkPipelineBindPoint bind_point) {
-    pipeline_bind_point = bind_point;
-    const auto lv_bind_point = ConvertToLvlBindPoint(bind_point);
-
-    LastBound &last_bound = cb_state.lastBound[lv_bind_point];
-    if (last_bound.pipeline_state) {
-        pipeline = last_bound.pipeline_state->VkHandle();
-        pipeline_layout = last_bound.pipeline_layout;
-        descriptor_sets.reserve(last_bound.per_set.size());
-        for (std::size_t i = 0; i < last_bound.per_set.size(); i++) {
-            const auto &bound_descriptor_set = last_bound.per_set[i].bound_descriptor_set;
-            if (bound_descriptor_set) {
-                descriptor_sets.push_back(std::make_pair(bound_descriptor_set->VkHandle(), static_cast<uint32_t>(i)));
-                if (bound_descriptor_set->IsPushDescriptor()) {
-                    push_descriptor_set_index = static_cast<uint32_t>(i);
-                }
-                dynamic_offsets.push_back(last_bound.per_set[i].dynamicOffsets);
-            }
-        }
-
-        if (last_bound.push_descriptor_set) {
-            push_descriptor_set_writes = last_bound.push_descriptor_set->GetWrites();
-        }
-        const auto &pipeline_layout = last_bound.pipeline_state->PipelineLayoutState();
-        if (pipeline_layout->push_constant_ranges == cb_state.push_constant_data_ranges) {
-            push_constants_data = cb_state.push_constant_data;
-            push_constants_ranges = pipeline_layout->push_constant_ranges;
-        }
-    } else {
-        assert(shader_objects.empty());
-        if (lv_bind_point == BindPoint_Graphics) {
-            shader_objects = last_bound.GetAllBoundGraphicsShaders();
-        } else if (lv_bind_point == BindPoint_Compute) {
-            auto compute_shader = last_bound.GetShaderState(ShaderObjectStage::COMPUTE);
-            if (compute_shader) {
-                shader_objects.emplace_back(compute_shader);
-            }
-        }
-    }
-}
-
-void RestorablePipelineState::Restore(VkCommandBuffer command_buffer) const {
-    if (pipeline != VK_NULL_HANDLE) {
-        DispatchCmdBindPipeline(command_buffer, pipeline_bind_point, pipeline);
-        if (!descriptor_sets.empty()) {
-            for (std::size_t i = 0; i < descriptor_sets.size(); i++) {
-                VkDescriptorSet descriptor_set = descriptor_sets[i].first;
-                if (descriptor_set != VK_NULL_HANDLE) {
-                    DispatchCmdBindDescriptorSets(command_buffer, pipeline_bind_point, pipeline_layout, descriptor_sets[i].second,
-                                                  1, &descriptor_set, static_cast<uint32_t>(dynamic_offsets[i].size()),
-                                                  dynamic_offsets[i].data());
-                }
-            }
-        }
-        if (!push_descriptor_set_writes.empty()) {
-            DispatchCmdPushDescriptorSetKHR(command_buffer, pipeline_bind_point, pipeline_layout, push_descriptor_set_index,
-                                            static_cast<uint32_t>(push_descriptor_set_writes.size()),
-                                            reinterpret_cast<const VkWriteDescriptorSet *>(push_descriptor_set_writes.data()));
-        }
-        if (!push_constants_data.empty()) {
-            for (const auto &push_constant_range : *push_constants_ranges) {
-                if (push_constant_range.size == 0) continue;
-                DispatchCmdPushConstants(command_buffer, pipeline_layout, push_constant_range.stageFlags,
-                                         push_constant_range.offset, push_constant_range.size, push_constants_data.data());
-            }
-        }
-    }
-    if (!shader_objects.empty()) {
-        std::vector<VkShaderStageFlagBits> stages;
-        std::vector<VkShaderEXT> shaders;
-        for (const vvl::ShaderObject *shader_obj : shader_objects) {
-            stages.emplace_back(shader_obj->create_info.stage);
-            shaders.emplace_back(shader_obj->VkHandle());
-        }
-        DispatchCmdBindShadersEXT(command_buffer, static_cast<uint32_t>(shader_objects.size()), stages.data(), shaders.data());
-    }
-}
-
-void CommandResources::Destroy(Validator &validator) {
-    if (instrumentation_desc_set != VK_NULL_HANDLE) {
-        validator.desc_set_manager->PutBackDescriptorSet(instrumentation_desc_pool, instrumentation_desc_set);
-        instrumentation_desc_set = VK_NULL_HANDLE;
-        instrumentation_desc_pool = VK_NULL_HANDLE;
-    }
-}
-
-void PreDrawResources::Destroy(Validator &validator) {
-    if (buffer_desc_set != VK_NULL_HANDLE) {
-        validator.desc_set_manager->PutBackDescriptorSet(desc_pool, buffer_desc_set);
-        buffer_desc_set = VK_NULL_HANDLE;
-        desc_pool = VK_NULL_HANDLE;
-    }
-
-    CommandResources::Destroy(validator);
-}
-
-void PreDispatchResources::Destroy(Validator &validator) {
-    if (indirect_buffer_desc_set != VK_NULL_HANDLE) {
-        validator.desc_set_manager->PutBackDescriptorSet(desc_pool, indirect_buffer_desc_set);
-        indirect_buffer_desc_set = VK_NULL_HANDLE;
-        desc_pool = VK_NULL_HANDLE;
-    }
-
-    CommandResources::Destroy(validator);
-}
-
-void PreTraceRaysResources::Destroy(Validator &validator) { CommandResources::Destroy(validator); }
-
-void PreCopyBufferToImageResources::Destroy(Validator &validator) {
-    if (desc_set != VK_NULL_HANDLE) {
-        validator.desc_set_manager->PutBackDescriptorSet(desc_pool, desc_set);
-        desc_set = VK_NULL_HANDLE;
-        desc_pool = VK_NULL_HANDLE;
-    }
-
-    if (copy_src_regions_buffer != VK_NULL_HANDLE) {
-        vmaDestroyBuffer(validator.vmaAllocator, copy_src_regions_buffer, copy_src_regions_allocation);
-        copy_src_regions_buffer = VK_NULL_HANDLE;
-        copy_src_regions_allocation = VK_NULL_HANDLE;
-    }
-
-    CommandResources::Destroy(validator);
 }
 
 }  // namespace gpuav
