+++ layers/gpu/core/gpuav_state_tracker.cpp
@@ -15,8 +15,8 @@
  * limitations under the License.
  */
 
-#include "gpu/gpu_shader_instrumentor.h"
-#include "gpu/gpu_state_tracker.h"
+#include "gpu/core/gpuav_state_tracker.h"
+#include "gpu/instrumentation/gpuav_shader_instrumentor.h"
 
 namespace gpu_tracker {
