--- layers/gpu/gpu_vuids.h
@@ -1,24 +1,0 @@
-/* Copyright (c) 2021-2024 The Khronos Group Inc.
- * Copyright (c) 2021-2024 Valve Corporation
- * Copyright (c) 2021-2024 LunarG, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include "gpu/gpu_validation.h"
-
-#pragma once
-// Getter function to provide kVUIDUndefined in case an invalid function is passed in
-namespace gpuav {
-const GpuVuid &GetGpuVuid(vvl::Func command);
-}  // namespace gpuav
