+++ layers/gpu/instrumentation/gpuav_shader_instrumentor.cpp
@@ -15,8 +15,9 @@
  * limitations under the License.
  */
 
-#include "gpu/gpu_shader_instrumentor.h"
-#include "gpu/gpu_state_tracker.h"
+#include "gpu/instrumentation/gpuav_shader_instrumentor.h"
+
+#include "gpu/core/gpuav_state_tracker.h"
 #include "chassis/chassis_modification_state.h"
 
 // Implementation for Descriptor Set Manager class
