+++ layers/gpu/instrumentation/gpuav_shader_instrumentor.h
@@ -16,8 +16,8 @@
  */
 #pragma once
 #include "generated/chassis.h"
-#include "gpu/gpu_resources.h"
-#include "gpu/gpu_state_tracker.h"
+#include "gpu/core/gpuav_state_tracker.h"
+#include "gpu/resources/gpuav_resources.h"
 #include "vma/vma.h"
 
 class DescriptorSetManager {
