+++ layers/gpu/resources/gpuav_descriptor_set.cpp
@@ -15,9 +15,10 @@
  * limitations under the License.
  */
 
-#include "gpu/gpu_descriptor_set.h"
-#include "gpu/gpu_validation.h"
-#include "gpu/gpu_subclasses.h"
+#include "gpu/resources/gpuav_descriptor_set.h"
+
+#include "gpu/core/gpuav.h"
+#include "gpu/resources/gpuav_subclasses.h"
 #include "gpu_shaders/gpu_shaders_constants.h"
 
 using vvl::DescriptorClass;
